Adjustes stats of weapons and added locale

+++ src/main/java/com/github/cassie365/edibleweapons/items/Quality.java
@@ -2,6 +2,6 @@ package com.github.cassie365.edibleweapons.items;
 
 public interface Quality {
 	public int getAttackBonus();
-	public int getSpeedBonus();
+	public float getSpeedBonus();
 
 }
